config: use AC_CONFIG_HEADERS to create a config.h file
This file contains C preprocessor #define statements which replace
the current -Ds added to each compilation invocation.

This makes the gcc output command easier to read and prevents exceeding
the max line limits on some computers. This is the preferred method
in al the xorg modules.

Note that cache.c already had #include config.h" which was incorrect
and misleading.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Reviewed-by: Arnaud Fontaine <arnau@debian.org>